Usuario casi terminado falta corregir creacion de Sala

+++ src/main/java/com/example/coworkingprojectback/repository/SalaRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface SalaRepository extends JpaRepository<Sala, Long> {
+public interface SalaRepository extends JpaRepository<Sala, Integer> {
     @Query("SELECT s FROM Sala s WHERE s.nombre = :nombre")
     Optional<Sala> buscarPorNombre(String nombre);
 }
